added encoder functions to embedding.py. Renaming of dim variables to be more descriptive.
